# swagger.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/plan/{id}": {
            "get": {
                "operationId": "PlanController_get",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "plan"
                ]
            },
            "put": {
                "operationId": "PlanController_put",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PlanDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "plan"
                ]
            },
            "delete": {
                "operationId": "PlanController_delete",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "plan"
                ]
            }
        },
...
        "/estado-plan": {
            "post": {
                "operationId": "EstadoPlanController_post",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EstadoPlanDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "estado-plan"
                ]
            },
            "get": {
                "operationId": "EstadoPlanController_getAll",
                "parameters": [
                    {
                        "name": "query",
                        "required": false,
                        "in": "query",
                        "description": "query - Filter. e.g. col1:v1,col2:v2 …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fields",
                        "required": false,
                        "in": "query",
                        "description": "fields - Fields returned. e.g. col1,col2 …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sortby",
                        "required": false,
                        "in": "query",
                        "description": "sortby - Sorted-by fields. e.g. col1,col2 …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order",
                        "required": false,
                        "in": "query",
                        "description": "order - Order corresponding to each sortby field, if single value, apply to all sortby fields. e.g. desc,asc …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "required": false,
                        "in": "query",
                        "description": "limit - Limit the size of result set. Must be an integer",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "offset",
                        "required": false,
                        "in": "query",
                        "description": "offset - Start position of result set. Must be an integer",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "populate",
                        "required": false,
                        "in": "query",
                        "description": "populate - show subqueries. Must be a boolean",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "estado-plan"
                ]
            }
        },
        "/estado-plan/{id}": {
            "get": {
                "operationId": "EstadoPlanController_get",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "estado-plan"
                ]
            },
            "put": {
                "operationId": "EstadoPlanController_put",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EstadoPlanDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "estado-plan"
                ]
            },
            "delete": {
                "operationId": "EstadoPlanController_delete",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "estado-plan"
                ]
            }
        },
        "/identificacion": {
            "post": {
                "operationId": "IdentificacionController_post",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IdentificacionDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "identificacion"
                ]
            },
            "get": {
                "operationId": "IdentificacionController_getAll",
                "parameters": [
                    {
                        "name": "query",
                        "required": false,
                        "in": "query",
                        "description": "query - Filter. e.g. col1:v1,col2:v2 …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fields",
                        "required": false,
                        "in": "query",
                        "description": "fields - Fields returned. e.g. col1,col2 …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sortby",
                        "required": false,
                        "in": "query",
                        "description": "sortby - Sorted-by fields. e.g. col1,col2 …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order",
                        "required": false,
                        "in": "query",
                        "description": "order - Order corresponding to each sortby field, if single value, apply to all sortby fields. e.g. desc,asc …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "required": false,
                        "in": "query",
                        "description": "limit - Limit the size of result set. Must be an integer",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "offset",
                        "required": false,
                        "in": "query",
                        "description": "offset - Start position of result set. Must be an integer",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "populate",
                        "required": false,
                        "in": "query",
                        "description": "populate - show subqueries. Must be a boolean",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "identificacion"
                ]
            }
        },
        "/identificacion/{id}": {
            "get": {
                "operationId": "IdentificacionController_get",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "identificacion"
                ]
            },
            "put": {
                "operationId": "IdentificacionController_put",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IdentificacionDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "identificacion"
                ]
            },
            "delete": {
                "operationId": "IdentificacionController_delete",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "identificacion"
                ]
            }
        },
...
    },
    "info": {
        "title": "plan_crud",
        "description": "API CRUD para el registro de planes para el cliente de planeacion",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "schemas": {
            "PlanDto": {
                "type": "object",
                "properties": {
                    "nombre": {
                        "type": "string"
                    },
                    "descripcion": {
                        "type": "string"
                    },
                    "tipo_plan_id": {
                        "type": "string"
                    },
                    "documento_id": {
                        "type": "string"
                    },
                    "formato": {
                        "type": "boolean"
                    },
                    "vigencia": {
                        "type": "string"
                    },
                    "dependencia_id": {
                        "type": "string"
                    },
                    "aplicativo_id": {
                        "type": "string"
                    },
                    "activo": {
                        "type": "boolean"
                    },
                    "estado_plan_id": {
                        "type": "string"
                    },
                    "padre_plan_id": {
                        "type": "string"
                    },
                    "arbol_padre_id": {
                        "type": "string"
                    },
                    "fecha_creacion": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "fecha_modificacion": {
                        "format": "date-time",
                        "type": "string"
                    }
                },
                "required": [
                    "nombre",
                    "descripcion",
                    "tipo_plan_id",
                    "documento_id",
                    "formato",
                    "vigencia",
                    "dependencia_id",
                    "aplicativo_id",
                    "activo",
                    "estado_plan_id",
                    "padre_plan_id",
                    "arbol_padre_id",
                    "fecha_creacion",
                    "fecha_modificacion"
                ]
            },
...
            "EstadoPlanDto": {
                "type": "object",
                "properties": {
                    "nombre": {
                        "type": "string"
                    },
                    "descripcion": {
                        "type": "string"
                    },
                    "codigo_abreviacion": {
                        "type": "string"
                    },
                    "activo": {
                        "type": "boolean"
                    },
                    "fecha_creacion": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "fecha_modificacion": {
                        "format": "date-time",
                        "type": "string"
                    }
                },
                "required": [
                    "nombre",
                    "descripcion",
                    "codigo_abreviacion",
                    "activo",
                    "fecha_creacion",
                    "fecha_modificacion"
                ]
            },
            "IdentificacionDto": {
                "type": "object",
                "properties": {
                    "nombre": {
                        "type": "string"
                    },
                    "descripcion": {
                        "type": "string"
                    },
                    "plan_id": {
                        "type": "string"
                    },
                    "dato": {
                        "type": "string"
                    },
                    "tipo_identificacion_id": {
                        "type": "string"
                    },
                    "activo": {
                        "type": "boolean"
                    },
                    "fecha_creacion": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "fecha_modificacion": {
                        "format": "date-time",
                        "type": "string"
                    }
                },
                "required": [
                    "nombre",
                    "descripcion",
                    "plan_id",
                    "dato",
                    "tipo_identificacion_id",
                    "activo",
                    "fecha_creacion",
                    "fecha_modificacion"
                ]
            },
...
        }
    }
}